tue, jun 13

+++ app/Http/Controllers/DepartmentController.php
@@ -6,17 +6,25 @@ use App\Models\Department;
 use GrahamCampbell\ResultType\Success;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class DepartmentController extends Controller
 {
     public function index()
     {
+        if (!Gate::allows('view-departments-lists')) {
+            return back()->with('error', 'Access denied!');
+        }
 
         return view('department.department', ['department' => Department::all()]);
     }
 
     public function create()
     {
+        if (!Gate::allows('create-departments')) {
+            return back()->with('error', 'Access denied!');
+        }
+
         request()->validate([
             'full_name' => ['required', 'string'],
             'short_name' => ['required', 'string'],
@@ -35,6 +43,10 @@ class DepartmentController extends Controller
 
     public function destroy(Department $department)
     {
+        if (!Gate::allows('destroy-departments')) {
+            return back()->with('error', 'Access denied!');
+        }
+
         $Success = $department->delete();
 
         if ($Success)
